Fixed refactor for get_schools
